killzone.css fixed
Finally the equivalent of GDscript code: extends Area2D

onready var timer = $Timer

func OnBodyEntered(body):
    print("You died!")
    timer.start()

func OnTimerTimeout():
    get_tree().reload_current_scene()
This code extends the Area2D node and initializes a timer variable with the Timer node ($Timer). It then defines two methods:

OnBodyEntered: This method is called when another physics body enters the area of the Area2D. It prints "You died!" to the console and starts the timer.
OnTimerTimeout: This method is called when the Timer node's timeout signal is emitted, indicating that the timer duration has passed. It reloads the current scene.
This GDScript code sets up a simple "kill zone" that triggers a reload of the scene when the player enters it and the specified timer duration passes.